+++ apple-codesign/src/code_resources.rs
@@ -1362,20 +1362,26 @@ impl CodeResourcesBuilder {
             // Unlike Apple's tooling, we recognize Mach-O binaries when the nested
             // flag isn't set and we automatically sign.
             //
-            // When we seal the file, we treat it as a regular file since the
-            // nested flag isn't set. Note that we need to read the signed/installed
-            // version of the file since signing will change its content.
-
-            let read_path = if crate::reader::path_is_macho(full_path)?
+            // Unless the path is marked for exclusion or shallow signing mode is
+            // active.
+            //
+            // The reason we exclude in shallow mode is that shallow mode is supposed
+            // to behave like Apple's `codesign` and that tool only signs the bundle's
+            // "main" Mach-O binary, not other binaries.
+            let sign_macho = crate::reader::path_is_macho(full_path)?
                 && !context
                     .settings
                     .path_exclusion_pattern_matches(root_rel_path)
-            {
+                && !context.settings.shallow();
+
+            let read_path = if sign_macho {
                 info!(
                     "non-nested file is a Mach-O binary; signing accordingly {}",
                     rel_path.display()
                 );
                 need_install = false;
+                // We need to read the signed/installed version of the file since
+                // signing will change its content.
                 context.sign_and_install_macho(full_path, rel_path)?.0
             } else {
                 info!("sealing regular file {}", rel_path_normalized);
@@ -1390,6 +1396,8 @@ impl CodeResourcesBuilder {
                 FilesFlavor::Rules2
             };
 
+            // When we seal the file, we treat it as a regular file since the
+            // nested flag isn't set.
             self.resources
                 .seal_regular_file(flavor, rel_path_normalized, digests, optional)?;
         }